Separate non-rpc behavior from KeybaseDaemonRPC
This cut and pastes non-RPC behavior from KeybaseDaemonRPC
into a KeybaseSericeBase (base) class.

No behavior should be changed by this.

This allows mobile to construct a non-RPC service based
implementation and talk directly to the client engine
package which implements all the keybase1 client interfaces,
 with all the behavior that KBFS expects from a
 KeybaseService like cacheing.
